+++ packages/api/src/utils/psa.js
@@ -1,5 +1,10 @@
+import { Validator } from '@cfworker/json-schema'
+import { PSAErrorInvalidData, PSAErrorRequiredData } from '../errors.js'
+import { normalizeCid } from '../utils/cid.js'
+
 /**
  * @typedef {'queued' | 'pinning' | 'failed' | 'pinned'} apiPinStatus
+ * @typedef {{ error?: PSAErrorInvalidData, data?: {} }} parsedData
  */
 
 /**
@@ -45,7 +50,7 @@ const psaStatusesToDBStatusesMap = {
 }
 
 /**
- * Maps a pinning api status array to db status accepted by the DB
+ * Maps a pinning api status array to db status accepted by the DB.
  * @param {string[]} statuses
  * @return {import('@web3-storage/db/postgres/pg-rest-api-types').definitions['pin']['status'][]}
  */
@@ -56,19 +61,192 @@ export const psaStatusesToDBStatuses = (statuses) => {
 }
 
 // Error messages
-// TODO: Refactor errors
 export const ERROR_CODE = 400
-export const ERROR_STATUS = 'INVALID_PIN_DATA'
-export const INVALID_CID = 'Invalid cid'
-export const INVALID_MATCH = 'Match should be a string (i.e. "exact", "iexact", "partial", "ipartial")'
-export const INVALID_META = 'Meta should be an object with string values'
-export const INVALID_NAME = 'Name should be a string'
-export const INVALID_ORIGINS = 'Origins should be an array of strings'
-export const INVALID_REQUEST_ID = 'Request id should be a string containing digits only'
-export const INVALID_REPLACE = 'Existing and replacement CID are the same'
-export const INVALID_STATUS = 'Status should be a list of "queued", "pinning", "pinned", or "failed"'
-export const INVALID_TIMESTAMP = 'Should be a valid timestamp'
-export const INVALID_LIMIT = 'Limit should be a number'
-export const REQUIRED_CID = 'CID is required'
-export const REQUIRED_REQUEST_ID = 'Request id is required'
-export const UNPERMITTED_MATCH = 'Match should be "exact", "iexact", "partial", or "ipartial"'
+export const DATA_NOT_FOUND = 'Requested data was not found.'
+export const INVALID_CID = 'The CID provided is invalid.'
+export const INVALID_REPLACE = 'Existing and replacement CID are the same.'
+export const INVALID_REQUEST_ID = 'Request id should be a string.'
+export const PINNING_FAILED = 'PSA_PINNING_FAILED'
+export const REQUIRED_REQUEST_ID = 'Request id is required.'
+
+export const DEFAULT_PIN_LISTING_LIMIT = 10
+export const MAX_PIN_LISTING_LIMIT = 1000
+
+// Validation schemas
+const listPinsValidator = new Validator({
+  type: 'object',
+  required: ['status'],
+  properties: {
+    name: { type: 'string', maxLength: 255 },
+    after: { type: 'string', format: 'date-time' },
+    before: { type: 'string', format: 'date-time' },
+    cid: { type: 'array', items: { type: 'string' } },
+    limit: { type: 'integer', minimum: 1, maximum: MAX_PIN_LISTING_LIMIT },
+    meta: { type: 'object' },
+    match: {
+      type: 'string',
+      enum: ['exact', 'iexact', 'ipartial', 'partial']
+    },
+    status: {
+      type: 'array',
+      items: {
+        type: 'string',
+        enum: ['queued', 'pinning', 'pinned', 'failed']
+      }
+    }
+  }
+})
+
+const postPinValidator = new Validator({
+  type: 'object',
+  required: ['cid'],
+  properties: {
+    cid: { type: 'string' },
+    name: { type: 'string', maxLength: 255 },
+    meta: { type: 'object', keys: { type: 'string' } },
+    origins: { type: 'array', items: { type: 'string' }, maxItems: 20 },
+    requestId: { type: 'string' }
+  }
+})
+
+/**
+ * Helper function to parse and validate payload for POST endpoint.
+ *
+ * @param {*} payload
+ * @returns parsedData
+ */
+export function transformAndValidate (payload) {
+  /** @type {*} */
+  const opts = {}
+  const {
+    cid,
+    name,
+    meta,
+    origins,
+    requestId
+  } = payload
+  let normalizedCid
+
+  // Validate CID.
+  if (cid) {
+    try {
+      normalizedCid = normalizeCid(cid)
+      opts.cid = cid
+    } catch (e) {
+      return {
+        error: new PSAErrorInvalidData(INVALID_CID),
+        data: undefined,
+        normalizeCid: undefined
+      }
+    }
+  }
+
+  if (name) opts.name = name
+  if (meta) opts.meta = meta
+  if (origins) opts.origins = origins
+  if (requestId) opts.requestId = requestId
+
+  const result = postPinValidator.validate(opts)
+
+  let data
+  let error
+
+  if (result.valid) {
+    data = opts
+  } else {
+    error = parseValidatorErrors(result.errors)
+  }
+
+  return { data, normalizedCid, error }
+}
+
+/**
+ * Helper function to parse and validate queryString for GET endpoint.
+ *
+ * @param {*} queryString
+ * @returns parsedData
+ */
+export function validateSearchParams (queryString) {
+  /** @type {*} */
+  const opts = {}
+  const {
+    after,
+    before,
+    cid,
+    limit,
+    match,
+    meta,
+    name,
+    status
+  } = queryString
+
+  // Validate CID(s).
+  if (cid) {
+    const cids = cid.split(',')
+    const sourceCids = []
+    for (const cid of cids) {
+      try {
+        normalizeCid(cid)
+        sourceCids.push(cid)
+      } catch (e) {
+        return {
+          error: new PSAErrorInvalidData(INVALID_CID),
+          data: undefined
+        }
+      }
+    }
+    opts.cid = sourceCids
+  }
+
+  if (limit) {
+    opts.limit = Number(limit)
+  } else {
+    opts.limit = DEFAULT_PIN_LISTING_LIMIT
+  }
+
+  if (name) opts.name = name
+  if (meta) opts.meta = meta
+  if (match) opts.match = match
+  if (before) opts.before = before
+  if (after) opts.after = after
+  if (status) opts.status = status.split(',')
+
+  const result = listPinsValidator.validate(opts)
+
+  let data
+  let error
+
+  if (result.valid) {
+    // Map statuses for DB compatibility.
+    opts.statuses = psaStatusesToDBStatuses(opts.status)
+    data = opts
+  } else {
+    error = parseValidatorErrors(result.errors)
+  }
+
+  return { data, error }
+}
+
+/**
+ * Helper function to parse error messages coming from the validator.
+ *
+ * @param {import('@cfworker/json-schema').OutputUnit[]} errors
+ * @returns {PSAErrorInvalidData}
+ */
+export function parseValidatorErrors (errors) {
+  // Pass through the error message of the last validation errors array.
+  // Last error message contains the most useful information.
+  const errorDetail = errors.pop()
+  if (!errorDetail) return new PSAErrorInvalidData()
+
+  const location = errorDetail.instanceLocation
+  const message = errorDetail.error
+  const errorType = errorDetail.keyword
+
+  switch (errorType) {
+    case 'required':
+      return new PSAErrorRequiredData(message)
+    default:
+      return new PSAErrorInvalidData(`${location}: ${message}`)
+  }
+}